fix: remove i18next built in def
