+++ packages/app-elements/src/locales/i18n.d.ts
@@ -1,0 +1,10 @@
+import type en from './en'
+
+declare module 'i18next' {
+  interface CustomTypeOptions {
+    defaultNS: 'translation'
+    resources: {
+      translation: typeof en
+    }
+  }
+}
